ipns: add wip krab integration
